Rename order resource
